fix: searchChannelsForListAction を commands に移動 (CQRS 違反修正)
DB upsert を含む searchChannelsForListAction が queries ファイルに
配置されていた CQRS 違反を修正。commands ファイルに移動し、
SearchChannelResult 型を queries から export して共有。
