+++ app/_actions/list-channel-commands.ts
@@ -5,6 +5,7 @@ import { createClient } from "@/lib/supabase/server";
 import { getUser } from "@/lib/auth";
 import { ApiError, handleApiError } from "@/lib/api/error";
 import { API_ERROR_CODES, type ApiResponse } from "@/lib/types/api";
+import type { SearchChannelResult } from "./list-channel-queries";
 
 /**
  * リストにチャンネルを追加
@@ -155,6 +156,100 @@ export async function removeChannelFromListAction(
       data: undefined,
     };
   } catch (err) {
+    return handleApiError(err);
+  }
+}
+
+/**
+ * チャンネルを検索（リスト追加用）
+ * YouTube APIから直接検索し、結果をデータベースに保存
+ */
+export async function searchChannelsForListAction(
+  query: string
+): Promise<ApiResponse<SearchChannelResult[]>> {
+  try {
+    const { searchChannels, getChannelDetails } =
+      await import("@/lib/youtube/api");
+    const youtubeResults = await searchChannels(query, 10);
+
+    if (youtubeResults.length === 0) {
+      return { success: true, data: [] };
+    }
+
+    const supabase = await createClient();
+    const searchResults: SearchChannelResult[] = [];
+
+    for (const ytChannel of youtubeResults) {
+      try {
+        const channelDetails = await getChannelDetails(
+          ytChannel.youtubeChannelId
+        );
+
+        const { data: upsertedChannel, error: upsertError } = await supabase
+          .from("channels")
+          .upsert(
+            {
+              youtube_channel_id: channelDetails.youtubeChannelId,
+              title: channelDetails.title,
+              description: channelDetails.description || null,
+              thumbnail_url: channelDetails.thumbnailUrl,
+              subscriber_count: channelDetails.subscriberCount || 0,
+              video_count: channelDetails.videoCount || 0,
+              view_count: channelDetails.viewCount || 0,
+              cache_updated_at: new Date().toISOString(),
+              updated_at: new Date().toISOString(),
+            },
+            { onConflict: "youtube_channel_id", ignoreDuplicates: false }
+          )
+          .select(
+            "id, youtube_channel_id, title, thumbnail_url, subscriber_count"
+          )
+          .single();
+
+        if (upsertError) {
+          console.error(
+            "[searchChannelsForListAction] Upsert error:",
+            ytChannel.youtubeChannelId,
+            upsertError
+          );
+          continue;
+        }
+
+        if (upsertedChannel) {
+          searchResults.push(upsertedChannel);
+        }
+      } catch (channelError) {
+        console.error(
+          "[searchChannelsForListAction] Error processing:",
+          ytChannel.youtubeChannelId,
+          channelError
+        );
+        continue;
+      }
+    }
+
+    return { success: true, data: searchResults };
+  } catch (err) {
+    console.error("[searchChannelsForListAction] Caught error:", err);
+
+    if (err && typeof err === "object" && "code" in err) {
+      const ytError = err as { code: string };
+      if (ytError.code === "QUOTA_EXCEEDED") {
+        return {
+          success: false,
+          error:
+            "YouTube APIのクォータを超過しました。しばらくしてから再度お試しください。",
+        };
+      }
+      if (ytError.code === "RATE_LIMIT") {
+        return {
+          success: false,
+          error:
+            "リクエスト制限に達しました。しばらくしてから再度お試しください。",
+        };
+      }
+    }
+
     return handleApiError(err);
   }
 }
