+++ app/_actions/list-channel.ts
@@ -1,12 +1,10 @@
 "use server";
 
 // CQRS分割: queries と commands から再エクスポート
-export {
-  getListChannelsAction,
-  searchChannelsForListAction,
-} from "./list-channel-queries";
+export { getListChannelsAction } from "./list-channel-queries";
 
 export {
   addChannelToListAction,
   removeChannelFromListAction,
+  searchChannelsForListAction,
 } from "./list-channel-commands";
